requirements.txt
gdown>=5.2.0
opencv-python>=4.9.0
numpy>=1.24.0
tqdm>=4.66.0
tensorflow>=2.16.0
matplotlib>=3.8.0
pillow>=10.3.0
keras>=3.0.0
roboflow
python-dotenv
inference_sdk

